solved user profile issue via setUserData(res.data) in createUser useEffect
