+++ src/context/userContext.jsx
@@ -4,9 +4,12 @@ import { useAuth0 } from "@auth0/auth0-react";
 import { Backend_URL } from "../BACKEND_URL.js";
 
 export const UserContext = createContext();
+const audience = process.env.REACT_APP_AUTH0_AUDIENCE;
+const scope = process.env.REACT_APP_AUTH0_SCOPE;
 
 export const UserContextProvider = (props) => {
-  const { user, isLoading, isAuthenticated } = useAuth0();
+  const { user, isLoading, isAuthenticated, getAccessTokenSilently } =
+    useAuth0();
   const [userData, setUserData] = useState([]);
   const [userEmail, setUserEmail] = useState("");
   const [allUserData, setAllUserData] = useState([]);
@@ -20,19 +23,38 @@ export const UserContextProvider = (props) => {
   }, [user, isAuthenticated, isLoading]);
 
   useEffect(() => {
-    if (userEmail) {
-      axios.get(`${Backend_URL}/users/${userEmail}`).then((response) => {
-        setUserData(response.data);
-      });
-    }
-    if (isUserDataUpdated === true) {
-      axios.get(`${Backend_URL}/users/${userEmail}`).then((response) => {
-        setUserData(response.data);
-      });
-    } else {
-      setIsUserDataUpdated(false);
-    }
-  }, [userEmail, isUserDataUpdated]);
+    const createUser = async () => {
+      if (isAuthenticated) {
+        const userInfo = {
+          username: user?.nickname,
+          first_name: user?.given_name || user?.name,
+          last_name: user?.family_name || "",
+          email_address: user?.email,
+          profile_pic_url: user?.picture,
+          voicevox_id: 1,
+        };
+        try {
+          // get access token
+          const accessToken = await getAccessTokenSilently({
+            audience: `${audience}`,
+            scope: `${scope}`,
+          });
+          await axios
+            .post(`${Backend_URL}/users/newUser`, userInfo, {
+              headers: {
+                Authorization: `Bearer ${accessToken}`,
+              },
+            })
+            .then((res) => {
+              setUserData(res.data);
+            });
+        } catch (err) {
+          console.log("Axios post to BE error", err);
+        }
+      }
+    };
+    createUser();
+  }, [getAccessTokenSilently, isAuthenticated, user]);
 
   useEffect(() => {
     if (!isLoading && isAuthenticated) {
@@ -45,6 +67,27 @@ export const UserContextProvider = (props) => {
       }
     }
   }, [isAuthenticated, isLoading]);
+
+  useEffect(() => {
+    const addOrUpdateUser = async () => {
+      if (userEmail) {
+        await axios
+          .get(`${Backend_URL}/users/${userEmail}`)
+          .then((response) => {
+            setUserData(response.data);
+          });
+      } else if (isUserDataUpdated === true) {
+        await axios
+          .get(`${Backend_URL}/users/${userEmail}`)
+          .then((response) => {
+            setUserData(response.data);
+          });
+      } else {
+        setIsUserDataUpdated(false);
+      }
+    };
+    addOrUpdateUser();
+  }, [userEmail, isUserDataUpdated]);
 
   return (
     <UserContext.Provider
