+++ src/pages/Home.jsx
@@ -2,7 +2,6 @@ import React from "react";
 import "./home.css";
 import { Link } from "react-router-dom";
 import GuideModalButton from "../components/GuideModalButton";
-import { LogoutButton } from "../components/Buttons";
 
 export default function Home() {
   return (
