+++ src/pages/Landing.jsx
@@ -28,41 +28,6 @@ export default function Landing() {
     return <Button onClick={() => loginWithRedirect()}>Log In</Button>;
   };
 
-  useEffect(() => {
-    const createUser = async () => {
-      if (isAuthenticated) {
-        console.log(isAuthenticated);
-        const userInfo = {
-          username: user?.nickname,
-          first_name: user?.given_name || user?.name,
-          last_name: user?.family_name || "",
-          email_address: user?.email,
-          profile_pic_url: user?.picture,
-          voicevox_id: 1,
-        };
-        try {
-          // get access token
-          const accessToken = await getAccessTokenSilently({
-            audience: `${audience}`,
-            scope: `${scope}`,
-          });
-          console.log(accessToken);
-          console.log(userInfo);
-          await axios.post(`${Backend_URL}/users/newUser`, userInfo, {
-            headers: {
-              Authorization: `Bearer ${accessToken}`,
-            },
-          });
-        } catch (err) {
-          console.log(userInfo);
-          console.log("Axios post to BE error", err);
-        }
-        console.log(userInfo);
-      }
-    };
-    createUser();
-  }, [getAccessTokenSilently, isAuthenticated, user]);
-
   const displayVoicevoxCharacters = voicevoxCharacters.map(
     (voicevoxCharacter, index) => (
       <img
